+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix issue with infinite loop upon setting `Circle#radius_x=` to a number with very small decimals
 - Fix issue with infinite loop upon setting `Circle#radius_y=` to a number with very small decimals
 - Fix issue with infinite loop upon setting `Circle#diameter=` to a number with very small decimals
+- Fix issue with infinite loop upon setting `Circle#width=` to a number with very small decimals
+- Fix issue with infinite loop upon setting `Circle#height=` to a number with very small decimals
 
 ## 1.0.0